Change ReplaySubject's to PublishSubjects
Since `self` was being sent to the subjects with a buffer of 1, it was
holding itself in the subjects indefinitely. The weakSelf I added in #30
only affects the closure itself, not the memory semantics of the subject
storing itself.

Looking back at RVMViewModel, they didn't have these as a replay
subject, so I changed them to publish subjects to match their semantics.

The throttling observable can keep it's subject since it's not sending
values of self.

I *think* this is correct, but please correct me on any of this. :)