Use shareReplayWhileConnected, tests compiling
I switched away from the manual subjects and just chain off of the
`activeObservable` with filtering. I *believe* unowned is ok here so
that we can leave the observable non-optional, since the only way the
`activeObservable` will fire a value is by something being set on
`self`, making that closure mutually dependent on self existing.

Removed the import of RxCocoa in the tests since it's not used anymore.
Tests still sporadically fail, I think it has something to do with the
interval timer, but not 100% sure.