Remove share behavior
Recommend we ditch trying to share the observable chain since nothing
terribly expensive is going on there, and having replay semantics causes
retain cycles if the consumer subscribes to the `didBecomeActive`
observable in the view model itself (and doesn't maintain the lifecycle
with `take(1)` or similar.) Either that, or we can move away from the
value sent being the view model. I don't have any instances where I
subscribe to the `didBecomeActive` observable after it's already active.